lesson_3/rag_bctc/requirements.txt
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0
qdrant-client==1.7.3
pypdf==4.1.0
python-dotenv==1.0.1
tiktoken==0.6.0
gradio==4.19.2
faiss-cpu==1.7.4 